Refactor(#44): 가독성을 위한 절대경로 수정
