+++ src/pages/meal-registration/index.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/base/colors.scss" as colors;
+@use "@styles/base/colors.scss" as colors;
 
 .tabWrapper {
   display: flex;
