+++ src/pages/mypage/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import Select from "../../components/select";
+import Select from "@components/select";
 
 export default function Mypage() {
   // 내 정보 수정, 등록된 사료/간식, 일간/월간, 즐겨찾기 -> 이렇게 네 개의 탭으로 나눌 예정.
